common.cpp: move PSA init from psa_util.cpp
